refactor: relocate soundroom feature files
