+++ src/features/soundroom/components/SoundRoom/HeaderBar.vue
