+++ src/features/soundroom/components/SoundRoom/MainCanvasStage/ListenerNode.vue
@@ -53,7 +53,7 @@
 
 <script setup>
 import { ref, onBeforeUnmount } from 'vue';
-import { useRoomStore } from '@/stores/useRoomStore';
+import { useRoomStore } from '@/features/soundroom/stores/useRoomStore';
 import { storeToRefs } from 'pinia';
