+++ src/features/soundroom/components/SoundRoom/SidebarLeft/ListenerReadout.vue
@@ -11,7 +11,7 @@
 
 <script setup>
 import { computed } from 'vue';
-import { useRoomStore } from '@/stores/useRoomStore';
+import { useRoomStore } from '@/features/soundroom/stores/useRoomStore';
 
 const store = useRoomStore()
