+++ src/features/soundroom/index.js
@@ -1,0 +1,3 @@
+export * from './components';
+export * from './composables';
+export * from './stores';
