+++ src/features/soundroom/stores/index.js
@@ -1,0 +1,1 @@
+export { useRoomStore } from './useRoomStore';
